chore: separate model
