+++ app/src/main/java/org/sopt/official/feature/attendance/model/AttendanceType.kt
@@ -1,0 +1,6 @@
+package org.sopt.official.feature.attendance.model
+
+enum class AttendanceType(val type: String) {
+    FIRST("1차 출석"),
+    SECOND("2차 출석")
+}
