+++ app/src/main/java/org/sopt/official/feature/attendance/model/state/TodayNoAttendanceCardState.kt
@@ -1,0 +1,7 @@
+package org.sopt.official.feature.attendance.model.state
+
+class TodayNoAttendanceCardState(
+    val eventDate: String,
+    val eventLocation: String,
+    val eventName: String,
+)
